Divide classes into respective packages
