+++ src/main/java/Blu/exception/BluException.java
@@ -1,0 +1,7 @@
+package Blu.exception;
+
+public class BluException extends Exception {
+    public BluException(String message) {
+        super("BluException:\n" + message);
+    }
+}
