+++ src/main/java/Blu/exception/IllegalCommandException.java
@@ -1,0 +1,7 @@
+package Blu.exception;
+
+public class IllegalCommandException extends BluException {
+    public IllegalCommandException(String message) {
+        super("Illegal Command Format: " + message);
+    }   
+}
