+++ src/main/java/Blu/exception/IllegalParameterException.java
@@ -1,0 +1,7 @@
+package Blu.exception;
+
+public class IllegalParameterException extends BluException {
+    public IllegalParameterException(String message) {
+        super("Illegal Parameter: " + message);
+    }
+}
