+++ src/main/java/Blu/exception/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package Blu.exception;
+
+public class InvalidCommandException extends BluException {
+    public InvalidCommandException(String message) {
+        super("Command not found: " + message);
+    }
+}
