+++ src/main/java/Blu/exception/StorageException.java
@@ -1,0 +1,8 @@
+package Blu.exception;
+
+public class StorageException extends BluException {
+    public StorageException(String message) {
+        super("Task Storage Exception: " + message);
+    }
+    
+}
