+++ src/main/java/Blu/utils/Messages.java
@@ -1,0 +1,8 @@
+package Blu.utils;
+
+public class Messages {
+    public static final String MESSAGE_GREET = "Hello! I'm Blu. What can I do for you?";
+    public static final String MESSAGE_STORAGE_PATH = "Using %s as storage file";
+    public static final String MESSAGE_EXIT = "Goodbye!";
+    public static final String MESSAGE_DIVIDER = "____________________________________________________________";
+}
