--- src/main/java/exception/BluException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class BluException extends Exception {
-    public BluException(String message) {
-        super("BluException:\n" + message);
-    }
-}
