--- src/main/java/exception/IllegalCommandException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class IllegalCommandException extends BluException {
-    public IllegalCommandException(String message) {
-        super("Illegal Command Format: " + message);
-    }   
-}
