--- src/main/java/exception/IllegalParameterException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class IllegalParameterException extends BluException {
-    public IllegalParameterException(String message) {
-        super("Illegal Parameter: " + message);
-    }
-}
