--- src/main/java/exception/InvalidCommandException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class InvalidCommandException extends BluException {
-    public InvalidCommandException(String message) {
-        super("Command not found: " + message);
-    }
-}
