--- src/main/java/exception/StorageException.java
@@ -1,8 +1,0 @@
-package exception;
-
-public class StorageException extends BluException {
-    public StorageException(String message) {
-        super("Task Storage Exception: " + message);
-    }
-    
-}
